[#12] refactor : 상수 클래스 object로 변경
