+++ coupon-platform-domain/src/main/kotlin/coupon_platform/domain/common/CommonConstants.kt
@@ -1,7 +1,5 @@
 package coupon_platform.domain.common
 
-class CommonConstants {
-    companion object {
-        const val EXTERNAL_ID_LENGTH = 12
-    }
+object CommonConstants {
+    const val EXTERNAL_ID_LENGTH = 12
 }
