+++ coupon-platform-domain/src/main/kotlin/coupon_platform/domain/issued_coupon/service/IssuedCouponStoreService.kt
@@ -1,10 +1,9 @@
 package coupon_platform.domain.issued_coupon.service
 
-import coupon_platform.domain.common.CommonConstants.Companion.EXTERNAL_ID_LENGTH
+import coupon_platform.domain.common.CommonConstants.EXTERNAL_ID_LENGTH
 import coupon_platform.domain.common.SuspendableRandomNumberGenerator
 import coupon_platform.domain.issued_coupon.dto.IssueCouponCommand
 import coupon_platform.domain.issued_coupon.repository.IssuedCouponStore
-import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
